0x04V2: Task10 Jinja2 Templates and Flask Script.
Added the required control structures to manage
which parts to be rendered based on the sent
parameters from the `9-states.py` python flask
script.

Added the two engines inside the flask script to
server the FileStorage and DBStorage.

Changed the data structure of the `state_dict`
variable to be a dictionary of dictionaries,
where `cities` key holds a dictionary value
data type. This helps in the sorting of cities
which was the issue we faced in the previous task
(task 8).

The cities are going to be sorted as
states through using `dictsort` built in function
in Jinja2.